Move home actions to merlins connector
