+++ web/app/containers/app/actions.js
@@ -25,7 +25,6 @@ import * as checkoutActions from '../../store/checkout/actions'
 import * as checkoutConfirmationActions from '../checkout-confirmation/actions'
 import * as checkoutShippingActions from '../../store/checkout/shipping/actions'
 import * as cartActions from '../../store/cart/actions'
-import * as homeActions from '../home/actions'
 import * as loginActions from '../login/actions'
 import * as productDetailsActions from '../product-details/actions'
 import * as footerActions from '../footer/actions'
@@ -142,9 +141,7 @@ export const fetchPage = (url, pageComponent, routeName, fetchUrl) => {
                 dispatch(receivedAction)
                 dispatch(process(receivedAction))
 
-                if (pageComponent === Home) {
-                    dispatch(homeActions.process(receivedAction))
-                } else if (pageComponent === Login) {
+                if (pageComponent === Login) {
                     dispatch(loginActions.process(receivedAction))
                 } else if (pageComponent === ProductDetails) {
                     dispatch(productDetailsActions.process(receivedAction))
