--- web/app/containers/home/actions.js
@@ -1,6 +1,0 @@
-import {createAction} from '../../utils/utils'
-import homeParser from './parsers/home'
-
-export const receiveData = createAction('Receive Home Page Data')
-
-export const process = ({payload: {$, $response}}) => receiveData(homeParser($, $response))
