+++ web/app/containers/home/container.js
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import * as commands from '../../integration-manager/commands'
+
 // Partials
 import HomeCarousel from './partials/home-carousel'
 import HomeCategories from './partials/home-categories'
@@ -11,6 +13,10 @@ const Home = () => {
             <HomeCategories />
         </div>
     )
+}
+
+Home.fetcher = (url, dispatch) => {
+    dispatch(commands.fetchHomeData(url))
 }
 
 export default Home
