+++ web/app/integration-manager/commands.js
@@ -4,13 +4,9 @@ export const register = (commands) => {
     connector = commands
 }
 
-export const fetchPdpData = (...args) => {
-    return connector.fetchPdpData(...args)
-}
+export const fetchPdpData = (...args) => connector.fetchPdpData(...args)
 
-export const addToCart = (...args) => {
-    return connector.addToCart(...args)
-}
+export const addToCart = (...args) => connector.addToCart(...args)
 
 export const fetchCheckoutShippingData = (...args) => connector.fetchCheckoutShippingData(...args)
 
@@ -19,3 +15,5 @@ export const submitShipping = (...args) => connector.submitShipping(...args)
 export const checkCustomerEmail = (...args) => connector.checkCustomerEmail(...args)
 
 export const submitSignIn = (...args) => connector.submitSignIn(...args)
+
+export const fetchHomeData = (...args) => connector.fetchHomeData(...args)
