+++ web/app/integration-manager/merlins-connector/home/parser.js
